feat: restructure chat message components and add media handling
